요구사항 3 -> controller 구현
