+++ web-application-server-gradle/src/main/java/controller/Controller.java
@@ -1,0 +1,8 @@
+package controller;
+
+import http.HttpRequest;
+import http.HttpResponse;
+
+public interface Controller {
+    void service(HttpRequest request, HttpResponse response);
+}
