+++ web-application-server-gradle/src/main/java/webserver/WebServer.java
@@ -6,7 +6,6 @@ import org.slf4j.LoggerFactory;
 import java.net.ServerSocket;
 import java.net.Socket;
 
-//test
 public class WebServer {
     private static final Logger log = LoggerFactory.getLogger(WebServer.class);
     private static final int DEFAULT_PORT = 8080;
